driver: pci: vpci overlay larger #interrupt-cells
Newer platforms use #interrupt-cells = <4> instead of <3> for GICv3.
Then the kernel fails to get the interrupts for ivshmem devices
when loading the root cell and parsing the generated vpci overlay:
  ivshmem-net 0005:00:01.0: of_irq_parse_pci: failed with rc=-22

Detect the size of the interrupt-cells and pad the generated
interrupt cells in the vpci overlay with <0> entries
like for other SPIs on these platforms.

Signed-off-by: Alex Zuepke <alex.zuepke@minervasys.tech>